+++ container-engine-lib/lib/docker_manager/docker_manager.go
@@ -728,8 +728,13 @@ func (manager *DockerManager) getContainerHostConfig(
 	manager.log.Debugf("Binds: %v", bindsList)
 
 	portMap := nat.PortMap{}
-	for containerPort := range exposedPorts {
-		portMap[containerPort] = nil
+	if shouldPublishAllPorts {
+		for containerPort := range exposedPorts {
+			portMap[containerPort] = []nat.PortBinding{
+				// Leaving this struct empty will cause Docker to automatically choose an interface IP & port on the host machine
+				{},
+			}
+		}
 	}
 
 	addedCapabilitiesSlice := []string{}
@@ -748,12 +753,14 @@ func (manager *DockerManager) getContainerHostConfig(
 		)
 	}
 
+	// NOTE: Do NOT use PublishAllPorts here!!!! This will work if a Dockerfile doesn't have an EXPOSE directive, but
+	//  if the Dockerfile *does* have an EXPOSE directive then _only_ the ports with EXPOSE will be published
+	// See also: https://www.ctl.io/developers/blog/post/docker-networking-rules/
 	containerHostConfigPtr := &container.HostConfig{
 		Binds: bindsList,
 		CapAdd: addedCapabilitiesSlice,
 		NetworkMode: container.NetworkMode(networkMode),
 		PortBindings: portMap,
-		PublishAllPorts: shouldPublishAllPorts,
 		ExtraHosts: extraHosts,
 	}
 	return containerHostConfigPtr, nil